float16

+++ oneat/NEATUtils/MovieCreator.py
@@ -566,7 +566,7 @@ yolo_v1 = True, yolo_v2 = False,  tshift  = 0, normalizeimage = True):
                          image = daskread(fname)[0]
                          print('raw', image.shape)
                          if normalizeimage:
-                            image = normalizeFloatZeroOne( image.compute().astype('float32'),1,99.8)
+                            image = normalizeFloatZeroOne( image.compute().astype('float16'),1,99.8)
                          segimage = daskread(Segfname)[0]
                          print('seg',segimage.shape)
                          for csvfname in filesCsv:
